Fixed a problem with trim_leading_brand_whitespace

+++ apparelrow/apparel/management/commands/trim_leading_brand_whitespace.py
@@ -20,14 +20,14 @@ class Command(BaseCommand):
         brand_names = set(Brand.objects.all().values_list('name', flat=True))
         print brands
         for brand in Brand.objects.iterator():
-            print brand
-            if brand.name.strip() in brands:
-                new_brand = brands[brand.name.strip()]
-                for p in brand.products.all():
-                    print 'Remapping %s to brand %s' % (p, new_brand)
-                    p.brand = new_brand
-                    p.save()
-                brand.delete()
+            if brand.name.strip() != brand.name:
+                if brand.name.strip() in brands:
+                    new_brand = brands[brand.name.strip()]
+                    for p in brand.products.all():
+                        print 'Remapping %s to brand %s' % (p, new_brand)
+                        p.brand = new_brand
+                        p.save()
+                    brand.delete()
             else:
                 brand.save()    # this will cause a strip of leading whitespace
